Update ccatp_sky_model.py
Change data type when reading healpy maps to np.float32

+++ ccatp_sky_model/ccatp_sky_model.py
@@ -526,11 +526,11 @@ def simulate_cmb(freq, cl_file = None, lensed = True, nside_out = 4096, lmax = N
         
             if lensed == True: 
         
-                CMB = hp.read_map('/vol/arc3/data1/sz/CCATp_sky_model/workspace_maude'+' CMB_lensed_CITA_mK')
+                CMB = hp.read_map('/vol/arc3/data1/sz/CCATp_sky_model/workspace_maude'+' CMB_lensed_CITA_mK', dtype = np.float32)
 
             else: 
             
-                CMB = hp.read_map('/vol/arc3/data1/sz/CCATp_sky_model/workspace_maude'+' CMB_unlensed_CITA_mK')     
+                CMB = hp.read_map('/vol/arc3/data1/sz/CCATp_sky_model/workspace_maude'+' CMB_unlensed_CITA_mK', dtype = np.float32)     
     
         if template == 'SO': 
 
@@ -549,7 +549,7 @@ def simulate_cmb(freq, cl_file = None, lensed = True, nside_out = 4096, lmax = N
 
                 file_name = '030_unlensedcmb_healpix.fits'
 
-        CMB = hp.read_map(data_path + file_name)
+        CMB = hp.read_map(data_path + file_name, dtype = np.float32)
 
     #Re-bin map if necessary
     if hp.get_nside(CMB) != nside_out:
@@ -617,14 +617,14 @@ def simulate_tSZ(freq, nside_out = 4096, lmax = None, beam_FWHM = None, template
         
         data_path='/vol/arc3/data1/sz/SO_sky_model/CMB_SZ_maps/'
         file_in ='tSZ_skymap_healpix_nopell_Nside4096_y_tSZrescale0p75.fits'
-        y_map = hp.read_map(data_path + file_in)
+        y_map = hp.read_map(data_path + file_in, dtype = np.float32)
 
         
     if template == 'CITA': 
         
         data_path='/vol/arc3/data1/sz/CITA/'
         file_in = 'tsz.fits'
-        y_map = hp.read_map(data_path + file_in)        
+        y_map = hp.read_map(data_path + file_in, dtype = np.float32)        
 
         
     if template == 'Sehgal': 
@@ -633,7 +633,7 @@ def simulate_tSZ(freq, nside_out = 4096, lmax = None, beam_FWHM = None, template
         file_in='030_tsz_healpix.fits'
         
         #Create Compton y-map : 
-        tSZ_30GHz = hp.read_map(data_path + file_in) * convert_units(freq, 1e-6, mjy2cmb=True)
+        tSZ_30GHz = hp.read_map(data_path + file_in, dtype = np.float32) * convert_units(freq, 1e-6, mjy2cmb=True)
            
         x_nu = np.array((h*30e9)/(k_B*T_CMB))    
         tSZ_SED = ((x_nu*(np.exp(x_nu)+1)/(np.exp(x_nu)-1))-4)
@@ -712,13 +712,13 @@ def simulate_kSZ(freq, nside_out = 4096, lmax = None, beam_FWHM = None, template
         
         data_path='/vol/arc3/data1/sz/SO_sky_model/CMB_SZ_maps/'
         file_in = '148_ksz_healpix_nopell_Nside4096_DeltaT_uK.fits'
-        kSZ = hp.read_map(data_path + file_in)
+        kSZ = hp.read_map(data_path + file_in, dtype = np.float32)
         
     if template == 'CITA':
         
         data_path = '/vol/arc3/data1/sz/CITA/'
         file_in = 'ksz.fits'
-        kSZ = hp.read_map(data_path + file_in)
+        kSZ = hp.read_map(data_path + file_in, dtype = np.float32)
         
     if template == 'Sehgal' : 
               
@@ -726,7 +726,7 @@ def simulate_kSZ(freq, nside_out = 4096, lmax = None, beam_FWHM = None, template
         file_in = '030_ksz_healpix.fits'  
 
         #Create compton-y map : 
-        kSZ = hp.read_map(data_path + file_in) * convert_units(freq, 1e-6, mjy2cmb=True)
+        kSZ = hp.read_map(data_path + file_in, dtype = np.float32) * convert_units(freq, 1e-6, mjy2cmb=True)
  
     #Re-bin map if necessary
     if hp.get_nside(kSZ) != nside_out:
